Add support for PRESERVE
Now the compression method can be changed without forcing a table
rewrite, by including the old method in the PRESERVE list.  For
supporting this the column will maintain the dependency with all
the supported compression methods.  So whenever the compression
method is altered the dependency is added with the new compression
method and the dependency is removed for all the old compression
methods which are not given in the preserve list.  If PRESERVE ALL
is given then all the dependency is maintained.

Dilip Kumar based on the patches from Ildus Kurbangaliev.
Design input from Tomas Vondra and Robert Haas
Reviewed by Robert Haas, Tomas Vondra, Alexander Korotkov and Justin Pryzby