Get datum from tuple which doesn't contain external data
Currently, we use HeapTupleGetDatum and HeapTupleHeaderGetDatum whenever
we need to convert a tuple to datum.  Introduce another version of these
routine HeapTupleGetRawDatum and HeapTupleHeaderGetRawDatum respectively
so that all the callers who doesn't expect any external data in the tuple
can call these new routines. Likewise similar treatment for
heap_copy_tuple_as_datum and PG_RETURN_HEAPTUPLEHEADER as well.  This is
the base patch for the optimization in the next patch therein we can try
to optimize the caller of the functions where we expect external varlena
by flattening any external toast pointer before forming the tuple.

Dilip Kumar based on the idea by Robert Haas