Alter table set compression
Add support for changing the compression method associated
with a column.  There are only built-in methods so we don't
need to rewrite the table, only the new tuples will be
compressed with the new compression method.

Dilip Kumar based on the patches from Ildus Kurbangaliev.
Design input from Robert Haas and Tomas Vondra.
Reviewed by Robert Haas, Tomas Vondra, Alexander Korotkov and Justin Pryzby