Invent HeapTupleGetRawDatum and friends.
HeapTupleGetRawDatum, HeapTupleHeaderGetRawDatum, and
PG_RETURN_HEAPTUPLEHEADER_RAW, and heap_copy_tuple_as_raw_datum are
just like HeapTupleGetDatum, HeapTupleHeaderGetDatum,
PG_RETURN_HEAPTUPLEHEADER, and heap_copy_tuple_as_datum, except that
they don't check that the returned tuple is free of external datums.
Instead, the caller must guarantee that this is the case.

The overarching idea here is that it's usually more efficient for the
caller to make sure to build the tuple originally without any external
datums, rather than building a tuple that may contain external datums
and then invoking a function which will possibly deform that tuple,
detoast those fields, and build a replacement tuple. Fortunately, very
few places are relying on this behavior of the above functions, which
means that this patch can safely change practically all the calls to
the existing functions to use the "raw" variants.

In the places that continue to use the non-raw variants after this
patch, there is an opportunity to improve performance by fixing them
so that they can. That work is planned for a follow-up patch.

That said, the main motivation for this patch is actually to cut down
on the number of places that depend on HeapTupleHasExternal tests to
avoid performance problems, and to make it clear where the remaining
places are. New features - such as configurable TOAST compression -
will not have the luxury of consuming an infomask bit to indicate
whether or not they are in use somewhere inside a tuple.

Dilip Kumar, based on an idea by me. Documentation changes also by me.

Discussion: http://postgr.es/m/CAFiTN-u3jrRDiuyGxGvNSGDXLhG4=U2orHnjk4B6WBy9Eo9kMQ@mail.gmail.com
Discussion: http://postgr.es/m/CAFiTN-s8SsWs2zCZL3TFxXwi4VE35J%3D4aKpduKVrTaL%2B3wnB0A%40mail.gmail.com