WAL logged CREATE DATABASE
Currently, CREATE DATABASE forces a checkpoint, then copies all the files,
then forces another checkpoint. The comments in the createdb() function
explain the reasons for this. The attached patch fixes this problem by making
create database completely WAL logged so that we can avoid the checkpoints.

This can also be useful for supporting the TDE, for example if we need different
encryption for the source and the target database then we can not reencrypt the
page data if we copy whole directory.  But with this patch we are copying page
by page so we have an opportunity to reencrypt the page before copying that to
the target database.