Refactor relmap load and relmap write functions
Currently, relmap reading and writing interfaces are tightly
coupled with shared_map and local_map of the database
it is connected to.  But as higher level patch set we need
interfaces where we can read relmap into any input memory
and while writing also we should be able to pass the map.

So as part of this patch, we are doing refactoring of the
existing code such that we can expose the read and write
interfaces that are independent of the shared_map and the
local_map, without changing any logic.

XXX For the code simplicity in write_relmap_file we are
updating the permanent memory copy outside the critical
section but we have already done the disk changes and it
is just a memory change so there is no reason for this
to be in the critical section.