Don't delay removing Tombstone file until next checkpoint
Currently, we can not remove the unused relfilenode until the
next checkpoint because if we remove them immediately then
there is a risk of reusing the same relfilenode for two
different relations during single checkpoint due to Oid
wraparound.

With this patch we are removing the need of keeping the
Tombstone files until the next checkpoint by making the
relfilenode unique within a cluster. For doing that we are
making RelFileNode.relNode 64 bits wider so that relfilenode
is never reused within a cluster.  But that will make buffer
tag wider by 32 bits so to avoid that we are removing the
ForkNumber from the buffer tag and using 8 high bits of
relNode for storing the fork number and remaining 56 bits
for the relfilenode.