Preliminary refactoring for supporting larger relfilenode
Currently, relNode in RelFileNode is Oid type and it can
wraparound so as part of the larger patch set we are trying
to make it 64 bit to avoid wraparound and that will make a
couple of other things simpler as explained in the next
patches.

So this is just a preliminary refactoring patch which will
replace the direct access of RelFileNode.relNode with a macros
to get and set the relNode without changing any other things.
And, the later patches will change the datatype of relNode
and hence the content of the macro to set and get the relNode.