Preliminary refactoring for storing fork number in RelFileNode
As described in the previous patch that we are planning to make
relNode as 64 bit but that will increase the size of the buffer
tag.  So in order to avoid that we are planning to store the
fork number in the first 8 bits of the relNode and the remaining
56 bits will be for the actual relNode.  So this is refactoring
for the same and as part of this patch we just avoid accessing
the fork number directly from the buffer tag and instead it will
be accessed using macro and the later patche will change the
macro definition such that it will access from the first 8 bits
of the relNode.