Change relfilenode from Oid to 56 bit wide int
As part of this patch change the RelFileNode.relNode 64 bit
wide.  Out of this high 8 bits will be reserved the fork number
and the remaining 56 bits will store the actual relation file.

So the idea is that if RelFileNode is used independently then
the RelFileNode.relNode will only store the 56 bits of the
relation and 8 high bits will be cleaned.  But as soon as we
store the RelFileNode as part of the BufferTag then we will
use these 8 high bits for the fork number.  So that we make the
relfilenode wider without actually increasing the size of
the buffer tag.