Preliminary refactoring for supporting larger relfilenode
Currently, relfilenode is Oid type and it can wrap around so as part of
the larger patch set we are trying to make it 64 bit to avoid wraparound
and that will make a couple of other things simpler as explained in the
next patches.

So this is just a preliminary refactoring patch as part of this, in
BufferTag, instead of keeping the RelFileNode, we will keep the
tablespace Oid, database Oid, and the relfilenode directly.  So that
once we change the relNode in RelFileNode to 64 bits the buffer tag
alignment padding will not change.