Allow ReadBufferWithoutRelcache to support unlogged relpersistence
At present, this function may only be used on permanent relations,
because we only use it during XLOG replay.  But now as part of the
bigger patch set, we will be using this for reading the buffer from
the database to which we are not connected.  So now we need this
for the unlogged relations as well.