Use 56 bits for relfilenode to avoid wraparound
As part of this patch, we will make the relfilenode 64 bits wide.
But the problem is that if we make it 64 bits wide then the size
of the BufferTag will be increased which will increase the memory
usage and that may also impact the performance.  So in order to
avoid that inside the buffer tag, instead of using 64 bits for the
relfilenode we will use 8 bits for the fork number and 56 bits for
the relfilenode.