Rename RelFileNode to RelFileLocator and relNode to RelNumber
Currently, the way relfilenode and relnode are used is really confusing.
Although there is some precedent for calling the number that pertains to
the file on disk "relnode" and that value when combined with the database
and tablespace OIDs "relfilenode," but it's definitely not the most obvious
thing, and this terminology is also not used uniformaly.

So as part of this patchset these variables are renamed to something more suited
with their usage. So the RelFileNode is renamed to the RelFileLocator
and all related variable declaration from relfilenode to relfilelocator.
And the relNode in the RelFileLocator is renamed to relNumber and along with that
the dbNode and spcNode are also renamed to dbOid and spcOid.  Along with that
all other references to relnode/relfilenode w.r.t to the ondisk file is renamed to
relnumber/relfilenumber.