Optimize subtrans slru
The small size of the SLRU buffer pools can sometimes become a performance problem, because it’s not difficult to have a workload where the number of buffers actively in use is larger than the fixed size buffer pool. However, just increasing the size of the buffer pool doesn’t necessarily help, because the linear search that we use for buffer replacement doesn’t scale, and also because contention on the single centralized lock limits scalability. To address these problems, I propose to introduce an optional SLRU buffer mapping hash table and apply it first to the Subtrans SLRU, which is a hot spot for workloads that make heavy use of subtransactions.