Optionally use LFU for buffer replacement in SLRU
In previous patch we have provided a support for optional
partitioned buffer mapping hash table.  With that we can
scale better by adding more SLRU buffer and also reduce
contention on a shared lock variable.  But we have noticed
in some cases the contention shifts to the shared lru counter
that is because multiple backend are modifying the same shared
variable and they are contending over frequent cache line miss.

In order to improve that we have implemented LFU algorithm where
it is enough to maintain the counter at per buffer level but we
do not need a shared most recently used counter.  There could be
other alternative approach to tackle this problem like modified
version of LRU algorithm where we do not need to update the
shared variable.