New catalog for parition ID to reloid mapping
TIDs alone are insufficient for uniquely identifying tuples
in global indexes because they include tuples from multiple
partitions. We might consider storing the relation OID along
with TID, as the combination of relation OID and TID is the
most straightforward way to uniquely identify a heap tuple.

However, this approach has its drawbacks. For instance, if a
partition is detached, the global index would not know to
ignore the tuples from that detached partition unless it
cleans out all tuples from that partition in the global index.

Therefore, we need an identifier for each partition that is
only valid while the partition is attached and becomes invalid
once the partition is detached. We call this identifier a
partition ID. However, when accessing tuples from the index,
we still need to convert the partition ID to the heap OID at
some point.

One might assume a 1-to-1 mapping between partition IDs and
relation OIDs, but this isn't feasible. In a multi-level partition
hierarchy, detaching a partitioned table from a higher-level
partitioned table invalidates the underlying leaf relation's
partition ID for the global indexes of the parent from which
it was detached. However, that partition ID should remain valid
for the lower-level partitions' global indexes where the leaf
partition is still attached. Therefore, for simplicity, we
maintain a partition ID for each global index and leaf relation
OID pair. This way, detaching a partition only requires
invalidating the specific global index and partition ID
combinations from which the leaf partition is being detached.

This patch provides a catalog for storing these mappings and a
cache for faster access. It also includes mechanisms for allocating
partition IDs and managing insertions and deletions in the catalog.