Provide Support for creating global index
This patch introduces support for creating global indexes in PostgreSQL,
providing a new syntax and mechanism to build the index.

Syntax: CREATE INDEX name ON table (column_list) GLOBAL;

TIDs alone are insufficient for uniquely identifying tuples in global indexes
because they include tuples from multiple partitions. To uniquely identify heap
tuples, we append a partition identifier. For detailed information, refer to
the comments in the patch 0001 commit message.

In this design, the partition-id is included as the last key column. The
rationale behind storing it as the last key column is that in various
scenarios, this column is treated as an extended index key column. For
example, in B-tree indexes, each index tuple must be uniquely identified.
If we encounter duplicate keys, we use heap TID as a tiebreaker. However,
for global indexes, relying solely on heap TID isn't adequate; we also
require the partition identifier. Including this as an additional key
column simplifies the process, as index tuples are arranged in key order,
and this column will automatically be part of that order.

Patch Details:

- Support for creating mappings into pg_index_partitions for each leaf
partition when a global index is created.
- Mechanism to build the B-tree global index from multiple partition data.
- Support for attaching and detaching partitions.