Provide support for global Index Scan Path
In previous patches we have added support for creating the global index.  Now
in this patch we provided a support in planner to choose a global index scan
and index only scan paths at for the append rel.

Currently we do not have support for selecting a bitmap scan using the global
index.  We may do that in future and if we need to do that we need to change
a executor such that we can build a sperate tidmap for each leaf relation while
scanning the global index and then do the bitmap heap scan partition at a time
based on the bitmap.

We also do not support the parallel index scan using the global index.  There
is nothing blocking as such but this is still a TODO.

Open Items
- ROWID_VAR is not cleanly handled in optimized, for now just made it workable but need to fix it properly
- In table_slot_callbacks(), now partiioned table can generate tuple by global index scan so we need proper
slot instead of just assigning a virtual slot.  This handling should be done maybe through AM callback?