Provide Support for creating global indexes and other  DDL  on global indexes
Syntax: CREATE INDEX name ON table (column_list) GLOBAL;

As described in commit message of the previous patch TIDs alone are
insufficient for uniquely identifying tuples in global indexes because
they include tuples from multiple partitions. To uniquely identify heap
tuples, we append a partitionID. For detailed information, refer to
the comments in the patch 0001 commit message.

In this design, the partitionID is included as the last key column of the
index. The rationale behind storing it as the last key column is that in
various scenarios, this column is treated as an extended index key column.
For example, in B-tree indexes, each index tuple must be uniquely identified.
If we encounter duplicate keys, we use heap TID as a tiebreaker. However,
for global indexes, relying solely on heap TID isn't adequate; we also
require the partition identifier. Including this as an additional key
column simplifies the process, as index tuples are arranged in key order,
and this column will automatically be part of that order.

Whenever a global index is created, this patch utilizes the interfaces from
the previous patch to assign a partitionID to each leaf relation of the
partitioned table on which the global index is being created. It then inserts
the (indexid, partitionID) -> reloid mapping into the pg_index_partition table
for each leaf relation.

When a global index is created on a partitioned table, a corresponding child
global index is created for each child relation. The top-level global index
is assigned the RELKIND_GLOBAL_INDEX relkind, while the child global indexes
are assigned RELKIND_GLOBAL_PARTITION_INDEX. Only the top-level global index
has physical storage; the child indexes exist as catalog entries. These child
indexes are used to directly access the top-level global index and the
partitioned relation where the global index is defined.  To facilitate this,
two additional fields were introduced in pg_index: indtoprelid and indtopindexid.
These fields eliminate the need to traverse the partition hierarchy when
operating directly on a child relation, as they provide quick access to the
ancestor's global index list.

Additionally, we need to create this mapping whenever a partition is attached.
Specifically, we must create mappings for all leaf partitions of the table being
attached, corresponding to the global indexes on the parent and all its ancestors
under which the new partition is being attached.

Similarly, when a partition is detached, we must invalidate the mappings for all
leaf partitions under the detached partition related to the global indexes on all
ancestors from which the partition is being detached.