Extension to validate stale page read
This extension validate stale page reads. The core idea is to record
the Log Sequence Number (LSN) of a page when it's evicted from memory.
Later, when that page is read back into the shared buffers, its current
LSN will be compared against the stored LSN to detect if the page content
is stale.

This extension register hooks while flushing the page and reading the
page.

TODO:
1. Make compatible with multiple versions by version compiler macros
2. Cleanup the hash when full
3. Performance test
4. Sanity check of the patch
5. Remove debug logs