Add relispublishable column to pg_class catalog
Previously, determining if a relation was eligible for logical
replication was performed dynamically via is_publishable_class().
This function relied on hard-coded OID checks and and relkind that
is performed at runtime.

This patch change by adding a "relispublishable" boolean column to
pg_class.  With this we don't need to check OID and relkind dynamically
we can just rely on ispublishable field.

Author: Dilip Kumar based on suggestion by Amit Kapila